// src/enum/feed.message.enum.ts
export enum FEED_MESSAGE {
    FETCHED_SUCCESS = 'Feed fetched successfully.',
    CREATED_SUCCESS = 'Feed created successfully.',
    UPDATED_SUCCESS = 'Feed updated successfully.',
    DELETED_SUCCESS = 'Feed deleted successfully.',
    NOT_FOUND = 'Feed not found.',
    INVALID_INPUT = 'Invalid input provided.'
}